requirements.txt
pandas==1.5.3
numpy==1.24.4
scikit-learn==1.2.2
faiss-cpu==1.7.4
tqdm==4.66.2
matplotlib==3.7.5
seaborn==0.12.2